+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Bug
 - \#168 - fix incorrect demo in readme
 
+## New Feature
+- \#170 - add index RNSG into HTTP handler
+- \#171 - add two new binary metric: SUBSTRUCTURE and SUPERSTRUCTURE
 
 # pymilvus 0.2.8(2020-03-11)