+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Bug
 - \#168 - fix incorrect demo in readme
+- \#172 - allow empty list in flush passing parameter
 
 ## New Feature
 - \#170 - add index RNSG into HTTP handler